+++ conf/app.conf.php
@@ -12,10 +12,6 @@ use Seablast\Seablast\SeablastConstant;
 return static function (SeablastConfiguration $SBConfig): void {
 //    $SBConfig->flag
 //        //->activate(SeablastConstant::FLAG_WEB_RUNNING)
-//        ->activate(AuthConstant::FLAG_USE_SOCIAL_LOGIN) // actual social login requires AuthApp:..social.._ID
-//        // - AuthApp:GOOGLE_CLIENT_ID
-//        // - AuthApp:FACEBOOK_APP_ID
-//        ->activate(AuthConstant::FLAG_REMEMBER_ME_COOKIE) // TODO check which default makes more sense
 //    ;
     $SBConfig
         ->setArrayString(I18nConstant::LANGUAGE_LIST, ['en', 'cs']) // default list of supported languages
@@ -26,5 +22,7 @@ return static function (SeablastConfiguration $SBConfig): void {
                 'model' => '\Seablast\I18n\Models\ApiLanguageModel',
             ]
         )
+        // Seablast::SeablastView uses this class for Latte filter translate
+        ->setString(SeablastConstant::TRANSLATE_CLASS, '\Seablast\I18n\SeablastTranslate') // since Seablast v0.2.7
     ;
 };