+++ CHANGELOG.md
@@ -19,13 +19,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### `Security` in case of vulnerabilities
 
-## [0.1] - 2025-07-20
+## [0.1] - 2025-07-27
 
 feat: library to handle language switching and localisation of selected strings
 
 ### Added
 
 - `/api/language` accesses `\Seablast\I18n\Models\ApiLanguageModel`
+- ULS langugage selector
+- localised items migration
 
 [Unreleased]: https://github.com/WorkOfStan/seablast-i18n/compare/v0.1...HEAD
 [0.1]: https://github.com/WorkOfStan/seablast-i18n/releases/tag/v0.1